requirements.txt
-i https://packages.idmod.org/api/pypi/pypi-production/simple
click~=7.1
idmtools~=1.6.0
idmtools-platform-comps~=1.6.0
idmtools-models~=1.6.0
emod-api~=1.16.5
matplotlib
astor
astunparse
bs4
keyring
